new commit

+++ assignment/a5.c
@@ -4,13 +4,13 @@ int main(){
 
     //printing arithmetic mean (a + b)/2
 
-    int a;
+    float a;
     printf("enter a"); 
-    scanf("%d", &a);
+    scanf("%f", &a);
 
-   int b;
+   float b;
  printf("enter b"); 
-    scanf("%d", &b);
+    scanf("%f", &b);
 
     printf("AM = %f", ((a+b) / 2));
     return 0;
